requirements.txt
firebase-admin>=6.5.0,<7
python-dotenv>=1.0.1,<2
fastapi>=1.8.0,<2
uvicorn[standard]>=0.15.0,<1
pyrebase4>=4.5.0,<5